requirements.txt
# 基本パッケージ
python-dotenv>=0.19.0  # 環境変数管理
requests>=2.26.0       # HTTP通信
beautifulsoup4>=4.9.3  # HTMLパース
lxml>=4.9.0           # XMLパース

# データ処理
pandas>=1.3.0         # データ分析
numpy>=1.21.0         # 数値計算
scikit-learn>=0.24.2  # 機械学習

# テスト・品質管理
pytest>=6.2.5         # テストフレームワーク
pytest-cov>=2.12.0    # カバレッジ測定
flake8>=3.9.0         # コードチェック
mypy>=0.910           # 型チェック

# ドキュメント
sphinx>=4.0.2         # ドキュメント生成
sphinx-rtd-theme>=0.5.2  # Sphinxテーマ

# ログ・モニタリング
loguru>=0.5.3         # ログ管理
